feat: re-initialized repository with svelte
The current implementation of my resume uses sproogen's
modern-resume-theme, which was quick to implement, but is not as
easy for me to modify as the JavaScript frameworks I have learned to use
over the past year. While I have no previous experience with Svelte, I
would like to learn, so I've decided to reverse engineer the
modern-resume-theme with Svelte. So, I have deleted all existing content
and replaced it with the svelte-ts template from Vite.