feat: added theme toggling
Currently, themes are imported in index.html, with media queries
controlling which theme (dark/light) is displayed. Because I have plans
to make the theme easily controllable by JavaScript in the future, I
have decided to do so now and avoid the possibility of a large
refactoring effort in the future. I have accomplished this by adding a
Theme.svelte component that changes the theme, based on a value
maintained in a Svelte store. This implementation retains the automatic
swapping of themes that normally occurs when a user changes their theme
preference in the browser, but also enables any component to change the
theme via the Svelte store.