refactor: separate about_me content from _config
Previously, the about_me content was stored in the main _config file,
which was only loaded when jekyll first started, so changes made could
only be viewed by restarting the jekyll server. As changes to this
content are frequent, I decided to move it to a separate yml file, from
which changes can be loaded automatically.