ci: remove extraneous artifact uploads
Previously, the `build` job would always upload an artifact, whether or
not it was to be deployed. Furthermore, when running the `deploy` job,
the artifact from the `build` job ran through an unnecessary,
intermediate pair of upload/download actions.

This commit removes the unnecessary pair of upload/download actions and
only uploads a pages artifact from the `build` job when the artifact is
to be deployed.