fix: reduce spacing below "about me" and "timeline"
This commit reduces the spacing below each of the "about me" and
"timeline" sections from mb-24 to mb-12. It also unifies the spacing
strategy to always use `margin-bottom` where possible, removing the
padding from the top of the "about me" and "timeline" elements and
adding equivalent margin to the bottom of the header.