feat!: update layout
This commit makes a dramatic change to the layout of sections, moving
the avatar from the "About Me" section to the right of the social links
in the header. It also joins a section's `title` and `name` with its
`description`, making the `date` display to the right of each section.

The purpose of this change is to make better use of the available
whitespace.