feat!: add drawer
This commit adds a second column to the right of the original content
(e.g. experience and education sections) at and above the `md`
breakpoint. At smaller breakpoints, this column defaults to hidden,
although it can be opened as a drawer through a new hamburger menu
button.

This new column/drawer is used to display a photo, contact information,
and a list of skills.