fix!: allow URIs without authority component
Previously, links were naively required to be provided as a so-called
`protocol` and `link` pair, where `protocol` was a `scheme` and `link`
referred the `authority` and `path`. The purpose of this format was to
provide a simple way to display urls without the `scheme`, while
including the scheme in `href`s by constructing them with
`{{protocol}}://{{link}}`.

This made the implicit assumption that all URIs had an authority
component, which was shown to be untrue by the presence of a `mailto`
link, which would incorrectly have a `//` placed between the `scheme`
(mailto) and the `path` (cjshearer@live.com).

This commit changes the format expected for urls provided in
`resume.yaml` back to a single string. Where the `scheme` is not desired
for display, the builtin URL parsing is used to assist removing it.

resolves #19