refactor!: move resume data to `/content` (#25)
* refactor: remove unnecessary titles from resume.yaml

* refactor: move resume data to /content

* refactor!: move entries to custom shortcode

This commit refactors the resume content in `_index.md` to use a custom
shortcode to render the main content of the resume. This isn't a true
refactor, as some minor spacing improvements between headings within the
main content of the resume have been made.

* build(deps): bump hugo to 0.120.3

This commit moves us from `hugo-bin-extended`, which hasn't had an
update to hugo in quite some time, to `hugo-bin`, which at the time of
writing has the latest hugo version. This update was needed to fix usage
of dynamic tailwind selectors containing characters like '&' and '>'
(see gohugoio/hugo#11372).

* fix: plain text URI display

This commit fixes plain text URI display, replacing the old approach of
parsing and reconstructing the URI with a regex replacement that removes
the `scheme ":" ["//" authority]` from the URI.