feat: add plain text output (#27)
This commit adds support for a plain text output
of the resume, which may be useful for submission
to an ATS system. It can be found at `/index.txt`.

Note that this hasn't been tested with any
particular ATS system, so I make no guarantees
that this will work, but it will *probably* be
better than a pdf print of the website.