chore: add `dev:tailscale` script
This commit adds support for using Tailscale to
serve the website to the rest of my tailnet.

The motivation for this change was to be able to
view the live site on my phone during development.

Note that while Hugo can serve locally with SSL
certs, this also uses Tailscale to generate
certificates, as it will also distribute them.