refactor!: move favicon site param to page param
This commit moves and renames
`.Site.params.faviconEmojiTemplate` to
`.Page.params.faviconText`. This both clarifies
that the parameter accepts any text - not
necessarily an emoji, as well as furthering a
wider effort to make this page fully configurable
from one location.

* refactor: move favicon template to head partial

We move the favicon svg template to the head
partial, since it will never be used anywhere
else.

* fix: bust favicon cache during development

While using `hugo server`, the first generated
favicon would cached and used for the duration. We
now name the generated resource with its md5 hash
to prevent this.