refactor!: nuke unnecessary features
I'm not even going to try to document what changed
here.

So much of this was built as a learning exercise,
but now I just want a simple, responsive,
print/ATS friendly resume. No more grids, no more
plain-text output; just simple HTML and CSS.

Software is an asset. Code is a liability.