feat: add theme as hugo module (#47)
This commit exposes the theme as a hugo module
that can be imported into other hugo websites. As
a part of this change, it moves the example of the
website to a standard `exampleSite`.

Another change that comes with this commit is the
transition of all linting and formatting
tools/checks away from node, towards tools that
are installed and used with nix. This is done
primarily because consumers of the theme will need
to download the same node dependencies as our repo
has, and I don't want there to be any extraneous
dependencies that are purely used for managing
this repository.

resolves #34