feat!: universalize
This is yet another commit with way too much stuff
in it.

The ultimate goal to make this a general hugo
theme that supports arbitrary content, upon which
I happen to build a resume and blog. I still need
to add a list template for handling the latter,
but I'm carving this off now to get some other
work done.

The site is now composed of a list of `head`
partials, a menu used for the header and footer,
between which the content of a page is rendered.
Most of these elements are, or will be
configurable or overridable from `hugo.toml`.

Along the way, I removed tailwindcss, as I don't
have a component heavy website layout and the
dependency adds more complexity than value.

I have also removed some of the components from
the main css and moved them into something
specific for my site (see
`exampleSite/assets/custom.css`).